requirements.txt
pymongo~=4.7.2
fastapi~=0.111.0
python-dotenv~=1.0.1
loguru~=0.7.2